refactor: split the component
