+++ resources/js/components/TwoFactorSetupModal.vue
@@ -1,0 +1,211 @@
+<script setup lang="ts">
+import InputError from '@/components/InputError.vue';
+import { Button } from '@/components/ui/button';
+import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle } from '@/components/ui/dialog';
+import { PinInput, PinInputGroup, PinInputSlot } from '@/components/ui/pin-input';
+import { useClipboard } from '@/composables/useClipboard';
+import { Form } from '@inertiajs/vue3';
+import { Check, Copy, Loader2, ScanLine } from 'lucide-vue-next';
+import { computed, nextTick, ref, watch } from 'vue';
+
+interface Props {
+    isOpen: boolean;
+    requiresConfirmation: boolean;
+    twoFactorEnabled: boolean;
+}
+
+interface Emits {
+    (e: 'update:isOpen', value: boolean): void;
+}
+
+const props = defineProps<Props>();
+const emit = defineEmits<Emits>();
+
+const { recentlyCopied, copyToClipboard } = useClipboard();
+
+const qrCodeSvg = ref<string | null>(null);
+const manualSetupKey = ref<string | null>(null);
+const hasSetupData = computed(() => {
+    return qrCodeSvg.value !== null && manualSetupKey.value !== null;
+});
+
+const showVerificationStep = ref(false);
+const code = ref<number[]>([]);
+const codeValue = computed(() => code.value.join(''));
+
+const pinInputContainerRef = ref<HTMLElement | null>(null);
+
+const modalConfig = computed(() => {
+    if (props.twoFactorEnabled) {
+        return {
+            title: 'You have enabled two factor authentication.',
+            description: 'Two factor authentication is now enabled, scan the QR code or enter the setup key in authenticator app.',
+            buttonText: 'Close',
+        };
+    }
+
+    if (showVerificationStep.value) {
+        return {
+            title: 'Verify Authentication Code',
+            description: 'Enter the 6-digit code from your authenticator app',
+            buttonText: 'Continue',
+        };
+    }
+
+    return {
+        title: 'Turn on 2-step Verification',
+        description: 'To finish enabling two factor authentication, scan the QR code or enter the setup key in authenticator app',
+        buttonText: 'Continue',
+    };
+});
+
+const handleModalNextStep = () => {
+    if (props.requiresConfirmation) {
+        showVerificationStep.value = true;
+        nextTick(() => {
+            pinInputContainerRef.value?.querySelector('input')?.focus();
+        });
+        return;
+    }
+
+    emit('update:isOpen', false);
+    return;
+};
+
+const fetchSetupData = async () => {
+    try {
+        const [qrResponse, keyResponse] = await Promise.all([
+            fetch(route('two-factor.qr-code'), { headers: { Accept: 'application/json' } }),
+            fetch(route('two-factor.secret-key'), { headers: { Accept: 'application/json' } }),
+        ]);
+
+        const { svg } = await qrResponse.json();
+        const { secretKey } = await keyResponse.json();
+
+        qrCodeSvg.value = svg;
+        manualSetupKey.value = secretKey;
+    } catch (error) {
+        console.error('Failed to fetch 2FA setup data:', error);
+    }
+};
+
+watch(
+    () => props.isOpen,
+    (isOpen) => {
+        if (isOpen && !qrCodeSvg.value) {
+            fetchSetupData();
+        }
+    },
+);
+
+defineExpose({ hasSetupData });
+</script>
+
+<template>
+    <Dialog :open="isOpen" @update:open="emit('update:isOpen', $event)">
+        <DialogContent class="sm:max-w-md">
+            <DialogHeader class="flex items-center justify-center">
+                <div class="mb-3 w-auto rounded-full border border-border bg-card p-0.5 shadow-sm">
+                    <div class="relative overflow-hidden rounded-full border border-border bg-muted p-2.5">
+                        <div class="absolute inset-0 grid grid-cols-5 opacity-50">
+                            <div v-for="i in 5" :key="`col-${i}`" class="border-r border-border last:border-r-0" />
+                        </div>
+                        <div class="absolute inset-0 grid grid-rows-5 opacity-50">
+                            <div v-for="i in 5" :key="`row-${i}`" class="border-b border-border last:border-b-0" />
+                        </div>
+                        <ScanLine class="relative z-20 size-6 text-foreground" />
+                    </div>
+                </div>
+                <DialogTitle>{{ modalConfig.title }}</DialogTitle>
+                <DialogDescription class="text-center">
+                    {{ modalConfig.description }}
+                </DialogDescription>
+            </DialogHeader>
+
+            <div class="relative flex w-auto flex-col items-center justify-center space-y-5">
+                <template v-if="!showVerificationStep">
+                    <div class="relative mx-auto flex max-w-md items-center overflow-hidden">
+                        <div class="relative mx-auto aspect-square w-64 overflow-hidden rounded-lg border border-border">
+                            <div
+                                v-if="!qrCodeSvg"
+                                class="absolute inset-0 z-10 flex aspect-square h-auto w-full animate-pulse items-center justify-center bg-background"
+                            >
+                                <Loader2 class="size-6 animate-spin" />
+                            </div>
+                            <div v-else class="relative z-10 overflow-hidden border p-5">
+                                <div v-html="qrCodeSvg" class="flex aspect-square size-full items-center justify-center" />
+                            </div>
+                        </div>
+                    </div>
+
+                    <div class="flex w-full items-center space-x-5">
+                        <Button class="w-full" @click="handleModalNextStep">
+                            {{ modalConfig.buttonText }}
+                        </Button>
+                    </div>
+
+                    <div class="relative flex w-full items-center justify-center">
+                        <div class="absolute inset-0 top-1/2 h-px w-full bg-border" />
+                        <span class="relative bg-card px-2 py-1">or, enter the code manually</span>
+                    </div>
+
+                    <div class="flex w-full items-center justify-center space-x-2">
+                        <div class="flex w-full items-stretch overflow-hidden rounded-xl border border-border">
+                            <div v-if="!manualSetupKey" class="flex h-full w-full items-center justify-center bg-muted p-3">
+                                <Loader2 class="size-4 animate-spin" />
+                            </div>
+                            <template v-else>
+                                <input type="text" readonly :value="manualSetupKey" class="h-full w-full bg-background p-3 text-foreground" />
+                                <button
+                                    @click="copyToClipboard(manualSetupKey || '')"
+                                    class="relative block h-auto border-l border-border px-3 hover:bg-muted"
+                                >
+                                    <Check v-if="recentlyCopied" class="w-4 text-green-500" />
+                                    <Copy v-else class="w-4" />
+                                </button>
+                            </template>
+                        </div>
+                    </div>
+                </template>
+
+                <template v-else>
+                    <Form
+                        :action="route('two-factor.confirm')"
+                        method="post"
+                        reset-on-error
+                        @error="code = []"
+                        @success="emit('update:isOpen', false)"
+                        v-slot="{ errors, processing }"
+                    >
+                        <input type="hidden" name="code" :value="codeValue" />
+                        <div ref="pinInputContainerRef" class="relative w-full space-y-3">
+                            <div class="flex w-full flex-col items-center justify-center space-y-3 py-2">
+                                <PinInput id="otp" placeholder="○" v-model="code" type="number" otp>
+                                    <PinInputGroup>
+                                        <PinInputSlot autofocus v-for="(id, index) in 6" :key="id" :index="index" :disabled="processing" />
+                                    </PinInputGroup>
+                                </PinInput>
+                                <InputError :message="errors?.confirmTwoFactorAuthentication?.code" />
+                            </div>
+
+                            <div class="flex w-full items-center space-x-5">
+                                <Button
+                                    type="button"
+                                    variant="outline"
+                                    class="w-auto flex-1"
+                                    @click="showVerificationStep = false"
+                                    :disabled="processing"
+                                >
+                                    Back
+                                </Button>
+                                <Button type="submit" class="w-auto flex-1" :disabled="processing || codeValue.length < 6">
+                                    {{ processing ? 'Confirming...' : 'Confirm' }}
+                                </Button>
+                            </div>
+                        </div>
+                    </Form>
+                </template>
+            </div>
+        </DialogContent>
+    </Dialog>
+</template>
